+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Customize background color, text color and font size on `RateDebug` screen data.
 - Enable/disable verbose screen data on `RateDebug`.
+- `RateManager.HasInstance` to check if the instance exists without creating it.
 ### Changed
 - Renamed `Debugger` to `RateDebug`.
 - Renamed `LogLevel` to `RateLogLevel`.